Fix lint errors in webSocketReducer test
- Auto-fixed formatting issues in `tests/unit/webSocketReducer.test.ts`.
